Janky list fix.
